6.01-open-api/pyproject.toml
[project]
name = "6-01-open-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "strands-agents-tools>=0.2.5",
    "strands-agents[openai]>=1.6.0",
]
